+++ canary_api_entrypoint.sh
@@ -1,0 +1,14 @@
+#!/usr/bin/env bash
+
+env >> /etc/environment \
+    && cron \
+    && gunicorn \
+        -b 0.0.0.0:2000 \
+        --workers 1 \
+        --log-level=DEBUG \
+        --timeout 30 \
+        --daemon \
+        -k gevent \
+        canarieapi.wsgi \
+    && service nginx start \
+    && /postgres-docker-entrypoint.sh postgres && "$@"